+++ tests/test_landings.py
@@ -6,16 +6,67 @@ import io
 import textwrap
 import unittest.mock as mock
 
+import pytest
+
 from landoapi.hg import AUTOFORMAT_COMMIT_MESSAGE, HgRepo
 from landoapi.models.landing_job import (
     LandingJob,
     LandingJobStatus,
     add_job_with_revisions,
 )
+from landoapi.models.revisions import Revision
 from landoapi.repos import SCM_LEVEL_3, Repo
 from landoapi.workers.landing_worker import LandingWorker
 
-PATCH_UNCHANGED = r"""
+
+@pytest.fixture
+def create_patch_revision(db):
+    """A fixture that fake uploads a patch"""
+
+    def _create_patch_revision(number, patch=PATCH_NORMAL_1):
+        revision = Revision()
+        revision.revision_id = number
+        revision.diff_id = number
+        revision.patch_bytes = patch.encode("utf-8")
+        db.session.add(revision)
+        db.session.commit()
+        return revision
+
+    return _create_patch_revision
+
+
+PATCH_NORMAL_1 = r"""
+# HG changeset patch
+# User Test User <test@example.com>
+# Date 0 0
+#      Thu Jan 01 00:00:00 1970 +0000
+# Diff Start Line 7
+add another file.
+diff --git a/test.txt b/test.txt
+--- a/test.txt
++++ b/test.txt
+@@ -1,1 +1,2 @@
+ TEST
++adding another line
+""".strip()
+
+PATCH_NORMAL_2 = r"""
+# HG changeset patch
+# User Test User <test@example.com>
+# Date 0 0
+#      Thu Jan 01 00:00:00 1970 +0000
+# Diff Start Line 7
+add another file.
+diff --git a/test.txt b/test.txt
+--- a/test.txt
++++ b/test.txt
+@@ -1,2 +1,3 @@
+ TEST
+ adding another line
++adding one more line
+""".strip()
+
+PATCH_NORMAL_3 = r"""
 # HG changeset patch
 # User Test User <test@example.com>
 # Date 0 0
@@ -453,7 +504,7 @@ def test_format_patch_success_unchanged(
 
     revisions = [
         create_patch_revision(1, patch=PATCH_FORMATTING_PATTERN_PASS),
-        create_patch_revision(2, patch=PATCH_UNCHANGED),
+        create_patch_revision(2, patch=PATCH_NORMAL_3),
     ]
     job_params = {
         "status": LandingJobStatus.IN_PROGRESS,
